refactor: Split auth function into separate module.
