+++ .gitignore
@@ -1,1 +1,3 @@
+# init
 /jaeseung/target/
+/minjun/target/
