+++ minjun/src/auth/login.rs
@@ -1,0 +1,14 @@
+#[derive(Debug)]
+pub struct LogInForm {
+    email: String,
+    password: String,
+}
+
+impl LogInForm {
+    pub fn new(email: String, password: String) -> LogInForm {
+        LogInForm {
+            email,
+            password,
+        }
+    }
+}
